:recycle: refactor: alterações na resposta para response e constantes
