+++ src/test/java/br/com/edu/ifmt/sacode/domain/controller/DespesaRestAdapterTest.java
@@ -12,7 +12,8 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import br.com.edu.ifmt.sacode.Controller.DespesaController;
+
+import br.com.edu.ifmt.sacode.application.DespesaRestAdapter;
 import br.com.edu.ifmt.sacode.domain.entities.Despesa;
 import br.com.edu.ifmt.sacode.domain.ports.LogPort;
 import br.com.edu.ifmt.sacode.domain.ports.DespesaPort;
@@ -39,7 +40,7 @@ import java.util.ResourceBundle;
 
 @SpringBootTest
 @AutoConfigureMockMvc
-public class DespesaControllerTest {
+public class DespesaRestAdapterTest {
     @Mock
     private DespesaPort despesaPortMock;
 
@@ -47,7 +48,7 @@ public class DespesaControllerTest {
     private LogPort logPortMock;
 
     @Mock
-    static private DespesaController despesaController;
+    static private DespesaRestAdapter despesaController;
 
     @Autowired
     private MockMvc mockMvc;
@@ -77,42 +78,4 @@ public class DespesaControllerTest {
     void buscarDespesasPorMesInvalido() throws Exception {
         mockMvc.perform(get("/despesas/2024-14")).andExpect(status().isBadRequest());
     }
-
-    /*
-     * @Test
-     * void shouldThrowException() {
-     * assertThrows(Exception.class, () -> {
-     * despesaController.buscarDespesasPorMes(null, 0, 0);
-     * });
-     * }
-     */
-
-    /*
-     * @Test
-     * public void testCriarDespesa_VerificarCampos() {
-     * Despesa despesa = mock(Despesa.class);
-     * assertThrows(DespesaException.class, () -> {
-     * despesaService.criarDespesa(despesa);
-     * });
-     * }
-     */
-    /*
-     * @Test
-     * public void testPorMes() throws DespesaException {
-     * UUID usuario = UUID.randomUUID();
-     * List<Despesa> despesas = new ArrayList<>();
-     * when(despesaPortMock.buscarDespesasPorPeriodo(usuario, LocalDate.of(2024, 4,
-     * 1), LocalDate.of(2024, 4, 30))).thenReturn(despesas);
-     * List<Despesa> result = despesaService.buscarDespesasPorMes(usuario, 2024, 4);
-     * assertEquals(despesas, result);
-     * }
-     */
-    /*
-     * @Test
-     * public void testPorMes_UserInvalido() {
-     * assertThrows(DespesaException.class, () -> {
-     * despesaService.buscarDespesasPorMes(null, 2024, 4);
-     * });
-     * }
-     */
 }
